requirements.txt
langchain-community
python-dotenv
huggingface-hub
Pillow
streamlit
requests 
langchain